refactor(bt): extract develop_fix_conditions submodule (closes #2604)
Split CheckIsVendorRoleNode and CheckCSFixNotYetReady out of develop_fix.py
into a new sibling module develop_fix_conditions.py before the 500-line cap
(BTND-07-004/BTND-07-006). Backward-compat re-exports added to develop_fix.py;
__init__.py updated to import conditions from the new home; architecture ratchet
test updated to reflect the VfdDimension constructor's new location.
